services/media-service/requirements.txt
fastapi
uvicorn
rembg
Pillow
